Reconciled two parallel deep-scan hardening waves. Semantic conflict resolutions
(not just textual):

- PBO CPCV-degenerate: ADOPTED phase-0's cpcv_exempt (proceed + distinct honest audit)
  over deepscan-wiring's BLOCK. CPCV is the default WF_MODE, so BLOCK-on-degenerate
  would strangle the entire pipeline. Fixed the producer→DB→consumer wiring end-to-end:
  pbo_degenerate_reason now rides INSIDE wf_metadata (persisted wholesale, read nested) —
  verified by gate-chain-integration pglite test. Rewrote finding1 test for the new policy.
- DLL-95 force-close: kept deepscan-wiring's awaited _confirmedForceClose (10s timeout —
  fixes the real hang risk) + grafted phase-0's completion audit, gated on a success
  boolean so it never falsely records "completed" when the close failed.
- paper-signal imports: unioned (forceCloseAllPositions + accountStrategyAssignments).
- style_c_handler intrabar-TP: auto-merged coherent (single price_reached, bar_high/low);
  unioned both sessions' test sets (C2 intrabar touch + Defect-2 no-MOVE_STOP_TO_BE).
- walk_forward WFE/BIF: my wfe_is_basis tag + their CPCV exemption coexist.
- AGENT-LOGS / System Map: both session entries kept; generated maps re-synced.

Verify: gate-chain-integration (35, pglite) + kill-switch (74) + PBO/BIF gates (62) +
their parity (87) + golden/cross-engine/static-C (113) + style_c_handler (36) GREEN;
tsc clean on merged files; all 3 CI gates GREEN (system-map driftItems=[]).

+++ src/server/__tests__/helpers/pglite-db.ts
@@ -122,7 +122,7 @@ CREATE TABLE IF NOT EXISTS backtests (
   avg_daily_pnl         NUMERIC,
   forge_score           NUMERIC,
   bif                   NUMERIC,
-  k_eff                 INTEGER,
+  k_eff                 NUMERIC,
   tier                  TEXT,
   equity_curve          JSONB,
   monthly_returns       JSONB,
